Add python script to rename images.
Given I forgot to change the folder in the camera
so now when I imported them they change and I want order,

So I'll make a script to rename the files to move it from
one range to another.
Just to push it to the end of my files in img folder.

Basically it should move the 0123_1 .. 0178_1 to the end in order.
It first detects them and then moves them.